[refactor] split cardano_sign_transaction_operation into v7/v8 part files
